eizz

+++ YoutubeToMp4/Program.cs
@@ -15,17 +15,9 @@ namespace YouTubeDownloader
             Directory.CreateDirectory(outputDirectory);
             Console.WriteLine("Started Program");
             // List of YouTube video URLs to download
-            List<string> videoUrls = new List<string>
-            {
-                "https://youtu.be/dHlpa5utQzQ?si=p9dfdsyYug3quKux",
-                "https://youtu.be/weZKm1kTrpc?si=RQnT4j9LiauyfAUP",
-                "https://youtu.be/kXMwZNRiPe0?si=QLE5J30SjXtyTwNS",
-                "https://youtu.be/9Zj0JOHJR-s?si=0_57adxXuFW4_0rx",
-                "https://youtu.be/EWjZOxs87yg?si=-TbodDxcSEStDB0w",
-                "https://youtu.be/rRzDNQPzbs4?si=kRRfgx_FQIhoWR_M",
-                "https://youtu.be/lCaun_EiJZQ?si=68xxVoxdNU3vgKH4"
-                // Add more video URLs as needed
-            };
+
+            List<string> videoUrls = GetUrls();
+
             Console.WriteLine("Urls:");
             foreach (var videoUrl in videoUrls)
             {
@@ -51,21 +43,43 @@ namespace YouTubeDownloader
             {
                 Console.WriteLine("An error occurred while downloading the videos: " + ex.Message);
             }
-            Console.WriteLine("-----------------");
-            Console.WriteLine("MP3 TIEM!!!");
-            //convert them all to mp3's
-            //delete this for loop if you don't want that
-            foreach (var video in exports)
+            Console.WriteLine("Do you want mp3 files? (y/n)");
+            string read = Console.ReadLine();
+            while(read != "y" && read != "n")
+            {
+                read = Console.ReadLine();
+            }
+
+            if(read == "y")
             {
                 Console.WriteLine("-----------------");
-                Console.WriteLine("Converting "+video);
-                convert.ConvertMedia(
-                    video,
-                    video+".mp3",
-                    "mp3");
-                File.Delete(video);
+                Console.WriteLine("MP3 TIEM!!!");
+                //convert them all to mp3's
+                //delete this for loop if you don't want that
+                foreach (var video in exports)
+                {
+                    Console.WriteLine("-----------------");
+                    Console.WriteLine("Converting "+video);
+                    convert.ConvertMedia(
+                        video,
+                        video+".mp3",
+                        "mp3");
+                    File.Delete(video);
+                }
             }
             Console.WriteLine("Done!");
+        }
+        static List<string> GetUrls(){
+            string[] lines = File.ReadAllLines(System.IO.Path.GetDirectoryName(System.Reflection.Assembly.GetEntryAssembly().Location)+"/urls.txt");
+
+            List<string> returnList = new List<string>();
+
+            foreach (var videoUrl in lines)
+            {
+                returnList.Add(videoUrl);
+            }
+
+            return returnList;
         }
         static async Task DownloadYouTubeVideo(string videoUrl, string outputDirectory)
         {
